invoice整理到type-guard
